Rename LG TV app to Jasper Veraart's Brilliant TV
