+++ .homeycompose/app.json
@@ -1,20 +1,20 @@
 {
-  "id": "com.lg.ir",
+  "id": "nl.jasperveraart.brilliant",
   "name": {
-    "en": "LG TV",
-    "nl": "LG TV"
+    "en": "Brilliant TV",
+    "nl": "Brilliant TV"
   },
   "version": "3.0.3",
   "sdk": 3,
   "compatibility": ">=5.0.0",
   "brandColor": "#AB234A",
   "author": {
-    "name": "Athom B.V.",
+    "name": "Jasper Veraart",
     "email": "support@athom.com"
   },
   "description": {
-    "en": "Control your LG TV using Infrared.",
-    "nl": "Bestuur je LG TV met Infrarood."
+    "en": "Control your TV using Infrared.",
+    "nl": "Bestuur je TV met Infrarood."
   },
   "category": "video",
   "images": {
