+++ README.md
@@ -1,4 +1,4 @@
-# LG TV
+# Brilliant TV
 
 This repository serves as inspiration for your own Homey app, to help you understand Homey Apps SDK concepts in a real-life context.
 
@@ -8,5 +8,5 @@ Read the [Homey Apps SDK Documentation](https://apps.developer.homey.app) for mo
 
 ## What does this app do?
 
-This app uses `homey-rfdriver` to enable Homey to control LG Televisions using Infrared.
+This app uses `homey-rfdriver` to enable Homey to control televisions using Infrared.
 Read more about Infrared on Homey at https://apps.developer.homey.app/wireless/infrared.
