+++ README.nl.txt
@@ -1,3 +1,3 @@
-Life's Good. En Life's Better als je Homey met je LG TV verbindt.
+Verbeter je Homey ervaring door je TV met deze app te verbinden.
 
 Homey kan je TV besturen via infrarood en je TV aan- en uitzetten, het kanaal wisselen en het volume instellen.
