+++ README.txt
@@ -1,3 +1,3 @@
-Life's Good. And Life's Better when you connect your LG TV with Homey.
+Enhance your Homey setup by connecting your TV with this app.
 
 Homey can control your TV with infrared and turn it on & off, change the channel and adjust the volume.
