+++ drivers/brilliant_tv/driver.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const RFDriverBrilliant = require('../../lib/RFDriverBrilliant');
+
+module.exports = class extends RFDriverBrilliant {
+
+};
