+++ drivers/brilliant_tv/pair/rf_ir_remote_add.html
