+++ drivers/brilliant_tv/pair/rf_ir_remote_learn.html
