--- drivers/lg_tv/driver.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const RFDriverLG = require('../../lib/RFDriverLG');
-
-module.exports = class extends RFDriverLG {
-
-};
