+++ lib/RFDeviceBrilliant.js
