+++ lib/RFDriverBrilliant.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const { RFDriver } = require('homey-rfdriver');
-const RFSignalLG = require('./RFSignalLG');
+const RFSignalBrilliant = require('./RFSignalBrilliant');
 
 module.exports = class extends RFDriver {
 
-  static SIGNAL = RFSignalLG;
+  static SIGNAL = RFSignalBrilliant;
 
 };
