+++ lib/RFSignalBrilliant.js
@@ -5,6 +5,6 @@ const { RFSignal } = require('homey-rfdriver');
 module.exports = class extends RFSignal {
 
   static FREQUENCY = 'ir';
-  static ID = 'lg';
+  static ID = 'brilliant';
 
 };
